patched security vulnerability & improved performance for opening links in a new tab
Before this commit, a link that opens in a new tab allowed the new tab access to the window. This meant a malicious page could hijack the page by navigating the user to a different URL. Furthermore, before this commit, any JavaScript executed on the client side in the new tab would affect our page's perforamance because the new page is on the same process. The security vulnerability was patched and the performance was improved by adding . To maintain browser compatibility, I added the 'noreferrer' value too. Read more about the security vulnerability and performance improvement here: https://developers.google.com/web/tools/lighthouse/audits/noopener